restructured project folders and changed name of components, pages, and css files
