+++ src/App.js
@@ -1,8 +1,8 @@
 import {Route, BrowserRouter as Router} from 'react-router-dom';
-import DashboardPage from './pages/dashboard-page';
+import DashboardPage from './pages/DashboardPage';
 import Login from './pages/Login.js';
 import Home from './pages/Home.js';
-import ProtectedRoute from './components/ProtectedRoute';
+import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute';
 // import './App.css';
 
 function App() {
@@ -10,7 +10,7 @@ function App() {
         <Router>
             <Route path='/' component={Home} exact />
             <Route path='/login' component={Login} />
-            <ProtectedRoute path='/dashboard-page' component={DashboardPage} />
+            <ProtectedRoute path='/dashboardPage' component={DashboardPage} />
         </Router>
     );
 }
