+++ src/styles/components/MovieSliderByGenre.module.css
