+++ src/styles/pages/Login.module.css
@@ -1,6 +1,6 @@
 .wallpaper {
     height: 100vh;
-    background: url('../img/login-wallpeper.jpg');
+    background: url('../../assets/img/login-wallpeper.jpg');
 }
 
 .formStyle {
